pyproject.toml
[tool.poetry]
name = "hw28_g"
version = "0.1.0"
description = ""
authors = ["emotzart <johnshuabratsk@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.1.3"
psycopg2-binary = "^2.9.5"
Pillow = "^9.3.0"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
